Move api implementations into IdentityService
