+++ src/beyond_identity/api/identities/patch.rs
@@ -1,13 +1,9 @@
 use super::{
     api::{IdentitiesApi, IdentityService},
-    types::{IdentitiesFieldName, Identity, PatchIdentityDetails},
+    types::{Identity, PatchIdentityDetails},
 };
-use crate::{
-    beyond_identity::api::common::{api_client::ApiClient, request::send_request, url::URLBuilder},
-    common::error::BiError,
-};
+use crate::common::error::BiError;
 use clap::Args;
-use http::Method;
 use serde::Serialize;
 
 // ===============================
@@ -17,37 +13,6 @@ use serde::Serialize;
 #[derive(Clone, Debug, Serialize)]
 pub struct PatchIdentityRequest {
     identity: PatchIdentityDetails,
-}
-
-// ===============================
-// API Function
-// ===============================
-
-pub async fn patch_identity(
-    service: &IdentityService,
-    identity_id: &str,
-    patch_request: &PatchIdentityRequest,
-) -> Result<Identity, BiError> {
-    let ApiClient {
-        config,
-        tenant_config,
-        client,
-    } = &service.api_client;
-    send_request(
-        client,
-        config,
-        tenant_config,
-        Method::PATCH,
-        &URLBuilder::build(tenant_config)
-            .api()
-            .add_tenant()
-            .add_realm()
-            .add_path(vec![IdentitiesFieldName::Identities.name(), identity_id])
-            .to_string()?,
-        Some(patch_request),
-    )
-    .await
-    .map(|details| Identity { identity: details })
 }
 
 // ===============================
